Refactor `removeSelectedReservations` in `DeleteReservationModal`
Add check to ensure that `deletePhysicalReservation` is only called if there is a physical material to be removed

+++ src/apps/reservation-list/modal/delete-reservation/delete-reservation-modal.tsx
@@ -34,19 +34,23 @@ const DeleteReservationModal: FC<DeleteReservationModalProps> = ({
       const reservationsToDelete = reservations
         .map((id) => Number(isFaust(id)))
         .filter((id) => id !== 0);
+
       const digitalMaterialsToDelete = reservations
         .map((id) => isIdentifier(id))
         .filter((id) => id !== null);
-      deletePhysicalReservation(
-        {
-          params: { reservationid: reservationsToDelete }
-        },
-        {
-          onSuccess: () => {
-            queryClient.invalidateQueries(getGetReservationsV2QueryKey());
+
+      if (reservationsToDelete.length > 0) {
+        deletePhysicalReservation(
+          {
+            params: { reservationid: reservationsToDelete }
+          },
+          {
+            onSuccess: () => {
+              queryClient.invalidateQueries(getGetReservationsV2QueryKey());
+            }
           }
-        }
-      );
+        );
+      }
 
       digitalMaterialsToDelete.forEach((id) =>
         deleteDigitalReservation(
@@ -60,6 +64,7 @@ const DeleteReservationModal: FC<DeleteReservationModalProps> = ({
           }
         )
       );
+
       close(modalId as string);
     }
   };
